fix: isolate connector tests into dedicated class
Connector test methods used a different model than other tests in the same class.
This made them flaky as connectors are stateful and potentially cached the previously
deployed model with a conflicting key. As after each test the engine is reset but connectors not.
